Ejercicio 2, Pedir nombre

+++ ex_02_03.py
@@ -1,8 +1,8 @@
 def main():
 	print("Ejercicio 3: Multiplicación")
-	xh = int(input("Enter hours: "))
-	xr = int(input("Enter rate: "))
-	xp = xh * xr
+	xh = input("Enter hours: ")
+	xr = input("Enter rate: ")
+	xp = float(xh) * float(xr)
 	print("Pay:", xp)
 	
 if __name__ == "__main__":
